avionics/rocket-checkout/pyproject.toml
[project]
name = "rocket-checkout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "synnax>=0.48.0",
    "termcolor>=3.2.0",
    "yaspin>=3.4.0",
]
